+++ walletAPI/services/interfaces/WalletTrackerdDBHandler.go
@@ -4,5 +4,5 @@ import "github.com/labora-wallet/walletAPI/model"
 
 type WalletTrackerDBHandler interface{
 	CreateWalletTracker(tracker model.WalletTrackerDTO) (int64, error)
-	GetWalletTrackByCustomerId(customerId int) (*[]model.WalletTracker, error)
+	GetWalletTrackByCustomerId(customerId int64) (*[]model.WalletTracker, error)
 }
